+++ ui/hooks/bridge/useEnableMissingNetwork.test.ts
@@ -1,5 +1,5 @@
-import { renderHookWithProvider } from '../../../../test/lib/render-helpers-navigate';
-import { createBridgeMockStore } from '../../../../test/data/bridge/mock-bridge-store';
+import { renderHookWithProvider } from '../../../test/lib/render-helpers-navigate';
+import { createBridgeMockStore } from '../../../test/data/bridge/mock-bridge-store';
 import * as ActionsModule from '../../store/actions';
 import { useEnableMissingNetwork } from './useEnableMissingNetwork';
